finish migrating functions to compile_state so bridge_macros absolute imports don't need to reference the crate builtins
