+++ builtins/src/types.rs
@@ -97,118 +97,15 @@
 //!  [`Value`]::CallFrame(Handle),
 //!  [`Value`]::Error(Handle),
 
-pub mod numbers;
-pub mod primitives;
-pub mod string_char;
-
-#[cfg(doc)]
-use bridge_types::{LooseString, SloshChar};
-use compile_state::state::SloshVm;
-use slvm::{VMResult, Value};
-
-pub trait SlFrom<T>: Sized {
-    /// Converts to this type from the input type.
-    fn sl_from(value: T, vm: &mut SloshVm) -> VMResult<Self>;
-}
-
-pub trait SlInto<T>: Sized {
-    /// Converts this type into the (usually inferred) input type.
-    fn sl_into(self, vm: &mut SloshVm) -> VMResult<T>;
-}
-
-impl<T, U> SlInto<U> for T
-where
-    U: SlFrom<T>,
-{
-    fn sl_into(self, vm: &mut SloshVm) -> VMResult<U> {
-        U::sl_from(self, vm)
-    }
-}
-
-pub trait SlFromRef<'a, T>: Sized
-where
-    Self: 'a,
-{
-    /// Converts to this type from the input type.
-    fn sl_from_ref(value: T, vm: &'a mut SloshVm) -> VMResult<Self>;
-}
-
-pub trait SlIntoRef<'a, T>: Sized
-where
-    T: 'a,
-{
-    /// Converts to this type from the input type.
-    fn sl_into_ref(self, vm: &'a mut SloshVm) -> VMResult<T>;
-}
-
-impl<'a, T, U> SlIntoRef<'a, U> for T
-where
-    U: SlFromRef<'a, T>,
-    U: 'a,
-{
-    fn sl_into_ref(self, vm: &'a mut SloshVm) -> VMResult<U> {
-        U::sl_from_ref(self, vm)
-    }
-}
-
-pub trait SlAsRef<'a, T: ?Sized> {
-    /// Converts this type into a shared reference of the (usually inferred) input type.
-    fn sl_as_ref(&self, vm: &'a mut SloshVm) -> VMResult<&'a T>;
-}
-
-// SlAsRef lifts over &
-impl<'a, T: ?Sized, U: ?Sized> SlAsRef<'a, U> for &'a T
-where
-    T: SlAsRef<'a, U>,
-{
-    #[inline]
-    fn sl_as_ref(&self, vm: &'a mut SloshVm) -> VMResult<&'a U> {
-        <T as SlAsRef<'a, U>>::sl_as_ref(*self, vm)
-    }
-}
-
-// SlAsRef lifts over &mut
-impl<'a, T: ?Sized, U: ?Sized> SlAsRef<'a, U> for &'a mut T
-where
-    T: SlAsRef<'a, U>,
-{
-    #[inline]
-    fn sl_as_ref(&self, vm: &'a mut SloshVm) -> VMResult<&'a U> {
-        <T as SlAsRef<'a, U>>::sl_as_ref(*self, vm)
-    }
-}
-
-pub trait SlAsMut<'a, T: ?Sized> {
-    /// Converts this type into a mutable reference of the (usually inferred) input type.
-    fn sl_as_mut(&mut self, vm: &'a mut SloshVm) -> VMResult<&'a mut T>;
-}
-
-// SlAsMut lifts over &mut
-impl<'a, T: ?Sized, U: ?Sized> SlAsMut<'a, U> for &'a mut T
-where
-    T: SlAsMut<'a, U>,
-{
-    #[inline]
-    fn sl_as_mut(&mut self, vm: &'a mut SloshVm) -> VMResult<&'a mut U> {
-        (*self).sl_as_mut(vm)
-    }
-}
-
-impl SlFrom<Value> for Value {
-    fn sl_from(value: Value, _vm: &mut SloshVm) -> VMResult<Self> {
-        Ok(value)
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    //#[test]
+    #[test]
     fn macro_passing_tests() {
         let t = trybuild::TestCases::new();
-        t.pass("trybuild/tests/*pass.rs");
+        t.pass("tests/*pass.rs");
     }
 
-    //#[test]
+    #[test]
     fn macro_failing_tests() {
         let t = trybuild::TestCases::new();
         t.compile_fail("trybuild/tests/*fail.rs");
