+++ compile_state/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 
 [dependencies]
 slvm = { path = "../vm" }
+bridge_types = { workspace = true }
