+++ compile_state/src/lib.rs
@@ -2,6 +2,7 @@ use crate::state::{CompileEnvironment, SloshVm, SloshVmTrait};
 use slvm::CallFuncSig;
 
 pub mod state;
+pub mod types;
 
 pub fn add_builtin(
     env: &mut SloshVm,
