+++ slosh/src/docs.rs
@@ -1,6 +1,6 @@
-use builtins::types::SlFrom;
 use compile_state::add_builtin;
 use compile_state::state::{SloshVm, SloshVmTrait};
+use compile_state::types::SlFrom;
 use lazy_static::lazy_static;
 use regex::{Regex, RegexBuilder};
 use slvm::VMErrorObj::Message;
